programs/oracle/programs/nexus-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "oracle-shared",
    "websockets>=12.0",
    "python-dotenv>=1.0",
    "anthropic>=0.30",
    "openai>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["nexus_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
